Added two py files to extend person and movie csv with Wikidata API calls on budget, death date etc.
There are two files in src/cleaner One named extension_person_wiki.py, which takes the in the Person.csv and output Person_extended.csv with birth_date, death_date, start_activity, end_activity if they were found in the wikidata -> we might put this to a boolean value True/False of whether the actor/director is still active. The dates are currently not in the right form, so choose either year or yyyy-mm-dd etc.

The second file is named extension_budget_wiki.py, which takes the in the Movie.csv and output Movie_extended.csv with budget and revenue if found on wikidata.

I tried to incorporate imdb data as well, but this did not contain any more information than just using wikidata for the first 50 rows of Person.csv.

Fuzzy is used to find names that are similar to the current name.

The names were unique, so addition of movie was not necessary in my opinion.